Split Precompiled and PrecompiledManager
