+++ transaction-executor/bcos-transaction-executor/precompiled/PrecompiledManager.cpp
@@ -1,49 +1,5 @@
 #include "PrecompiledManager.h"
-#include "bcos-utilities/Overloaded.h"
 #include <memory>
-
-struct UnsupportedPrecompiledException : public bcos::Exception
-{
-};
-
-constexpr static void releaseOutput(const struct evmc_result* result)
-{
-    delete[] result->output_data;
-}
-
-bcos::transaction_executor::EVMCResult bcos::transaction_executor::Precompiled::call(
-    evmc_message const& message) const
-{
-    return std::visit(
-        bcos::overloaded{
-            [&](executor::PrecompiledContract const& contract) {
-                auto [success, output] = contract.execute({message.input_data, message.input_size});
-                auto gas = contract.cost({message.input_data, message.input_size});
-
-                auto buffer = std::unique_ptr<uint8_t>(new uint8_t[output.size()]);
-                std::copy(output.begin(), output.end(), buffer.get());
-                EVMCResult result{evmc_result{
-                    .status_code =
-                        (evmc_status_code)(int32_t)(success ? protocol::TransactionStatus::None :
-                                                              protocol::TransactionStatus::
-                                                                  RevertInstruction),
-                    .gas_left = message.gas - gas.template convert_to<int64_t>(),
-                    .gas_refund = 0,
-                    .output_data = buffer.release(),
-                    .output_size = output.size(),
-                    .release = std::addressof(releaseOutput),
-                    .create_address = {},
-                    .padding = {},
-                }};
-
-                return result;
-            },
-            [](std::shared_ptr<precompiled::Precompiled> const&) {
-                BOOST_THROW_EXCEPTION(UnsupportedPrecompiledException());
-                return EVMCResult{evmc_result{}};
-            }},
-        m_precompiled);
-}
 
 bcos::transaction_executor::PrecompiledManager::PrecompiledManager()
 {
