+++ transaction-executor/bcos-transaction-executor/precompiled/PrecompiledManager.h
@@ -1,24 +1,11 @@
 #pragma once
 
-#include "../Common.h"
-#include "bcos-executor/src/vm/Precompiled.h"
+#include "Precompiled.h"
 #include "bcos-utilities/FixedBytes.h"
 #include <utility>
 
 namespace bcos::transaction_executor
 {
-
-class Precompiled
-{
-private:
-    std::variant<executor::PrecompiledContract, std::shared_ptr<precompiled::Precompiled>>
-        m_precompiled;
-
-public:
-    Precompiled(decltype(m_precompiled) precompiled) : m_precompiled(std::move(precompiled)) {}
-
-    EVMCResult call(evmc_message const& message) const;
-};
 
 class PrecompiledManager
 {
